requirements.txt
flask
pyyaml
PyGithub